requirements.txt
Flask==0.10.1
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
SQLAlchemy==0.9.8
Werkzeug==0.10.1
alembic==0.7.4
argparse==1.2.1
arrow==0.5.4
docopt==0.6.2
docutils==0.12
itsdangerous==0.24
lockfile==0.10.2
pyjade==3.0.0
python-daemon==2.0.5
python-dateutil==2.4.0
requests==2.5.3
six==1.9.0
slacker==0.5.5
wsgiref==0.1.2
